remove unused style class, data methods, comp prop
