+++ client/src/components/TopToolbar.vue
@@ -1,23 +1,21 @@
 <template>
-<v-card
-    color="grey lighten-4"
+  <v-card
     flat
     tile
   >
     <v-toolbar
       dense
       color="primary"
-      class="toolB"
-      :height="dynamicToolbarHeight"
+      height="48"
     >  
-      <v-btn icon class="toolBtn">
-        <v-icon :size="dynamicToolbarIcon" class="toolBIcon" color="#fff">{{routeIcons}}</v-icon>
+      <v-btn icon>
+        <v-icon size="24" color="#fff">{{routeIcons}}</v-icon>
       </v-btn>
-      <v-toolbar-title color="#fff" class="whiteText toolBRouteText">{{$route.name}}</v-toolbar-title>
+      <v-toolbar-title class="whiteText ml-n3">{{$route.name}}</v-toolbar-title>
 
       <v-spacer></v-spacer>
 
-      <v-toolbar-title color="#fff" class="whiteText toolBBarText">Barcode Gen</v-toolbar-title>
+      <v-toolbar-title class="whiteText">Barcode Gen</v-toolbar-title>
     </v-toolbar>
   </v-card>
 </template>
@@ -41,35 +39,6 @@ export default class TopToolbar extends Vue {
     if (this.$route.name === "Create") return "mdi-barcode"
     if (this.$route.name === "Print") return "mdi-printer"
   }
-
-  get dynamicToolbarHeight() {
-    if (this.window.width >= 3840) return "100"
-    if (this.window.width >= 3000 && this.window.width <= 3839) return "75"
-    return "48"
-  }
-
-  get dynamicToolbarIcon() {
-    if (this.window.width >= 3840) return "80"
-    if (this.window.width >= 3000 && this.window.width <= 3839) return "60"
-    return "24"
-  }
-
-  // Lifecycle Events ----------------
-  // ---------------------------------
-  created() {
-    window.addEventListener('resize', this.handleResize)
-    this.handleResize();
-  }
-  destroyed() {
-    window.removeEventListener('resize', this.handleResize)
-  }
-
-  // Methods -------------------------
-  // ---------------------------------
-  handleResize() {
-    this.window.width = window.innerWidth;
-    this.window.height = window.innerHeight;
-  }
 }
 </script>
 
@@ -77,34 +46,4 @@ export default class TopToolbar extends Vue {
 .whiteText {
   color: #fff !important;
 }
-
-/* @media screen and (max-width: 4096px) and (min-width: 3001px) and (min-height: 2000px) and (max-height: 2304px) {
-  .toolBIcon {
-    margin-left: 80px !important;
-  }
-  .toolBBarText, .toolBRouteText {
-    font-size: 45px;
-  }
-  .toolBRouteText {
-    margin-left: 60px;
-  }
-  .toolBBarText {
-    margin-right: 30px;
-  }
-}
-
-@media screen and (max-width: 3000px) and (min-width: 2881px) and (min-height: 1600px) and (max-height: 2000px) {
-  .toolBIcon {
-    margin-left: 80px !important;
-  }
-  .toolBBarText, .toolBRouteText {
-    font-size: 45px;
-  }
-  .toolBRouteText {
-    margin-left: 60px;
-  }
-  .toolBBarText {
-    margin-right: 30px;
-  }
-} */
 </style>
